wrong id check

+++ src/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PrismaService } from '../prisma/prisma.service';
@@ -26,10 +26,24 @@ export class PostsService {
 
   async findOne(id: number) {
     const post = await this.prisma.post.findUnique({ where: { id } });
+    if (!post) {
+      throw new HttpException(
+        'Post with such id doent exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     return post;
   }
 
   async update(id: number, updatePostDto: UpdatePostDto) {
+    const post = await this.prisma.post.findUnique({ where: { id } });
+    if (!post) {
+      throw new HttpException(
+        'Post with such id doent exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     const result = await this.prisma.post.update({
       where: { id },
       data: updatePostDto,
@@ -38,6 +52,14 @@ export class PostsService {
   }
 
   async remove(id: number) {
+    const post = await this.prisma.post.findUnique({ where: { id } });
+    if (!post) {
+      throw new HttpException(
+        'Post with such id doent exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     const result = await this.prisma.post.delete({ where: { id } });
     return result;
   }
