+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { RegisterDto } from '../auth/dto/register.dto';
@@ -19,10 +19,24 @@ export class UsersService {
 
   async findOne(id: number) {
     const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) {
+      throw new HttpException(
+        'User with such id doent exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     return user;
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) {
+      throw new HttpException(
+        'User with such id doent exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     const result = await this.prisma.user.update({
       where: { id },
       data: updateUserDto,
@@ -31,6 +45,14 @@ export class UsersService {
   }
 
   async remove(id: number) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) {
+      throw new HttpException(
+        'User with such id doent exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     const result = await this.prisma.user.delete({ where: { id } });
     return result;
   }
